First commit for the chess project
THis is the first initial commit for the chess project. Contians the full game, as it stands now. It inlcudes the base chess game, with all the basic rules applied. It also includes the ability to see your possible moves for a certain piece at the spot it is now. WIll add a menu of sorts, and the ability to castle later, and more