More messaging stuff
You can now send messages to other people, and it will be recorded in the database, and it will show up in their messages when they look. Also, when you click on an email, it updates the database to show that the email has been read, and I did the same for deletion. I'm now woking on implementing an auto completion feature that will show the user possible employees that include the name that the user has typed in, so that the user can send the email to the proper employee, so that the spelling is correct, and so that it will get to them.